scripts/requirements.txt
requests==2.32.3
supabase==2.9.1
python-dotenv==1.0.1
rapidfuzz==3.10.1
